coverity: ensure lchan->si->* bitmask sizes
osmo-bts' gsm_lchan->si bitmasks are 32 bit wide, and loops use
_MAX_SYSINFO_TYPE to iterate them. Make sure that _MAX_SYSINFO_TYPE
(from libosmocore) never surpasses 32 without us noticing the problem.

(However unlikely that is.)

Related: CID#465523

+++ tests/misc/misc_test.c
@@ -145,7 +145,7 @@ static void test_sacch_get(void)
 
 	/* initialize the input. */
 	for (i = 1; i < _MAX_SYSINFO_TYPE; ++i) {
-		lchan.si.valid |= (1 << i);
+		lchan.si.valid |= (((uint32_t)1) << i);
 		memset(GSM_LCHAN_SI(&lchan, i), i, GSM_MACBLOCK_LEN);
 	}
